### 2023-05-16 #5468 @editorialbot
[PRE REVIEW]: CRSocket: A web app component facilitating the administration of digital trials from a separate device
**Submitting author:** @henrikdvn (<a href="http://orcid.org/0000-0002-8515-1670">Henrik Dvergsdal</a>)
**Repository:** https://github.com/henrikdvn/CRSocket
**Branch with paper.md** (empty if default branch):
**Version:** 1.13 (CommTest)
