### 2023-05-16 #5468 @oliviaguest
@henrikdvn can you try putting the paper at the top level in the repo? I think this causes the problem with generating the pdf?
