+++ paper.md
@@ -16,7 +16,7 @@ affiliation: 1
 affiliations:
 - name: Nord University, Norway
 index: 1
-date: 29 December 2022
+date: 16 May 2023
 bibliography: paper.bib
 ---
 
