### 2023-07-07 #5468 @britta-wstnr
Hi @rabdill and @Milantex👋

Would you be willing to review the submission:
CRSocket: A web app component facilitating the administration of digital trials from a separate device
for [The Journal of Open Source Software (JOSS)](https://joss.theoj.org/)?

The review concerns the software [CRSocket](https://github.com/henrikdvn/CRSocket) and a [short paper](https://github.com/openjournals/joss-papers/blob/joss.05468/joss.05468/10.21105.joss.05468.pdf) about it. The review happens fully on GitHub.

You can learn more about the the review process and our conflict of interest policy in the reviewer guidelines [here](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html).

If you are available to review this submission, please let me know and I can add you as a reviewer. The review process will start (in a separate GitHub issue) once a sufficient number of reviewers have been found.

Thank you!
Britta
