### 2023-07-14 #5658 @editorialbot
[REVIEW]: CRSocket: A web app component facilitating the administration of digital trials from a separate device
**Submitting author:** @henrikdvn (<a href="http://orcid.org/0000-0002-8515-1670">Henrik Dvergsdal</a>)
**Repository:** https://github.com/henrikdvn/CRSocket
**Branch with paper.md** (empty if default branch):
**Version:** 1.13 (CommTest)
