### 2023-07-14 #5658 @rabdill
**The short version:** This application uses PHP 7.4, which is [no longer supported](https://www.php.net/supported-versions.php). Is that an issue, @britta-wstnr?

**The longer version:**
This is interesting work and I'm making my way through the review checklist, but I wanted to raise a question about the use of PHP: The [documentation](https://henrikdvn.github.io/CRSocket/general/Installation.html) currently states that CRSocket can run on "any web server that supports PHP 7.4 and common Apache HTTP functionality... Migration to PHP 8 will require some minor adjustments."

PHP 7.4 left active support [in 2021](https://www.php.net/supported-versions.php), and, as of last November, is no longer receiving even critical security patches. I don't see any journal policy about this, but I'm unsure of whether it's a good idea to publish a software paper that relies on a platform that is explicitly no longer supported, particularly when it is intended to be the foundation of larger applications that 1) will probably take a little while to develop, and 2) will probably be used for months or years at a time. So it's a _little_ outdated right now, but it will be considerably farther removed by the time someone integrates CRSocket into their application and deploys it for a study.

On the other hand, [most online PHP applications](https://www.zend.com/blog/planning-php-7-4-eol) are still using PHP 7.4, so this is a common issue—indeed, using a secure and supported version of PHP seems like the more unusual situation. So sure, it isn't best practice, but it's not rare, and there's no requirement that CRSocket applications actually communicate over the internet instead of a local network.

Looking at past reviews, I can't find many examples of situations like this: [A 2020 review](https://github.com/openjournals/joss-reviews/issues/2605#issuecomment-703921123) noted a similarly unsupported version of PHP, but that wasn't the deciding factor in its eventual rejection. A reviewer noted that a 2017 submission was [using Python 2.7](https://github.com/openjournals/joss-reviews/issues/440#issuecomment-351871064); the submission was accepted, but Python 2 was not yet end-of-life. This was also raised in [another 2019 review](https://github.com/openjournals/joss-reviews/issues/1663#issuecomment-539212089), but Python 2 was not EOL then either, and the issues raised were unrelated.

Anyway, I'm sorry if this is the wrong format for a question like this, but I didn't want to open an issue in the submitting repository for updating PHP if this was just an editorial call that I'm not supposed to be involved in.
