### 2023-07-14 henrikdvn/CRSocket#1 @rabdill
JOSS Review: Documentation
Hello! I'm one of the reviewers for this submission to JOSS, working on https://github.com/openjournals/joss-reviews/issues/5658. I don't mean to spam you with issues and comments, but I tried to split up my feedback into relevant chunks for discussion and will be submitting them as they come up. This issue has my notes about the "Documentation" section of the [JOSS reviewer checklist](https://joss.readthedocs.io/en/latest/review_checklist.html). (One minor note: The repository's readme currently has two links to documentation, one at `https://henrikdvn.github.io/CRSocket/` and one at `https://neurolab.no/docs/crsocket/`, which are slightly different. I reviewed the version at the first URL, which I believe is connected directly to the repository.)

>Automated tests: Are there automated tests or manual steps described so that the functionality of the software can be verified?

The "CommTest" app appears to be a thorough demonstration of CRSocket's functionality, with lots of visibility into what's going on internally and [some documentation](https://henrikdvn.github.io/CRSocket/apps/CommTest.html) of what is being displayed. I think this would work fine for someone tinkering with the code who wanted to verify they hadn't broken anything, but I don't believe it's adequately documented to do this. I'm confident a CRSocket expert could use this for all kinds of things, but the controller and responder dashboards both include dozens of esoteric phrases without a description in the docs—the toggles for "Log," "Buf" and "Itc," for example.

If the CommTest app is meant to validate the functionality of the software, I believe it needs at least some of these additions to make it fit for purpose:

- An explanation of what all of the sections and line items are documenting.
- Instructions for initializing the application
- Recommended interactions to exercise different functionality
- Expected state. For example, on loading the Commtest Responder from the project homepage, the "receiverState" field is listed as "closed," and the "senderState" field is blank. I assume this is expected, but I don't actually know. It would be very helpful to have a list of actions and how this information is expected to change in response.

It's fine by me if the CommTest app isn't actually intended to be for testing the software, but in that case, another approach for validating functionality would be needed.

>Community guidelines: Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support

As far as I can tell, this is currently missing from the repository and documentation website. For what it's worth, GitHub made [an example `CONTRIBUTING.md` document](https://github.com/github/docs/blob/5a827cbdaa69a709dd2ab62ae0cdf0dbaca240b3/CONTRIBUTING.md) that may be helpful.
