### 2023-07-14 henrikdvn/CRSocket#2 @rabdill
JOSS review: Software paper
Another issue documenting progress on https://github.com/openjournals/joss-reviews/issues/5658, this one about the [reviewer checklist](https://joss.readthedocs.io/en/latest/review_checklist.html) items regarding the manuscript.

>Summary: Has a clear description of the high-level functionality and purpose of the software for a diverse, non-specialist audience been provided?

The only shortcoming of the presented summary is that it doesn't indicate what an administrator would need to do to implement CRSocket in their experimental protocol. Phrases like "event-driven client/server web app infrastructure" are technically descriptive, but I believe it excludes a "non-specialist audience" who may need this tool but are not web developers. As it's currently described, it's not obvious to me that this is not "plug and play" software that someone could use without advanced technical skills. That's fine, but I think it would be helpful if this were more clearly explained.

>State of the field: Do the authors describe how this software compares to other commonly-used packages?

The paper has a "Statement of need" section that describes the project's origins and issues it aims to resolve ("potential sources of distraction," "make communication problems visible for the tester," etc.). However, there is no discussion of its relation to other work—how are people accomplishing this task now? Why is this new architecture an improvement, and in what ways? It's helpful to know that the code includes a "basic authentication scheme," for example, but is the implication that existing options do not?

>References: Is the list of references complete, and is everything cited appropriately that should be cited (e.g., papers, datasets, software)? Do references in the text use the proper [citation syntax](https://pandoc.org/MANUAL.html#extension-citations)?

This is a very minor note, but in places where multiple references are provided for a single statement, the formatting of these citations is slightly off: Each reference is given its own entry, like this:
```
many types of data collection [@Frank2016], [@Semmelmann2016], [@Kanerva2019], but the
```

The effect in the compiled version is that there are three separate parentheticals separated by commas:

>many types of data collection (Frank et al., 2016), (Semmelmann et al., 2016), (Kanerva et al., 2019), but the

If the references were formatted as specified, this shouldn't happen. I believe this is the correct setup:

```
[@Frank2016; @Semmelmann2016; @Kanerva2019]
```
