### 2023-07-16 #5658 @henrikdvn
Regarding the PHP version: There was an issue with version 8.0 at some point, but it now seems to work well with 8.2.5 which is the latest supported by my ISP. Haven't tested extensively, and I don't know about all versions in between, but I have now taken the chance to state that it supports "7.4 or later".
