### 2023-07-16 henrikdvn/CRSocket#1 @henrikdvn
Thanks for engaging with this, and feel free to give me all types of feedback. I will try to sort things out as fast as I can.

Regarding documentation differences: I have been working on some updates (mostly internal crosslinks and formatting) and wanted to complete them before committing. Anyways they should be in sync now.

I see your point regarding the CommTest app. I tried to compress the layout as much as possible so that it would be reasonably usable on mobile devices. Will try to incorporate your suggestions into the documentation, add some links from keywords in the app to other parts of the docs and possibly group different parts of the app into meaningful sections.

Regarding community guidelines: Thanks for the example. I'm not an expert on this. Will take a look at it and provide some guidelines.
