### 2023-07-16 henrikdvn/CRSocket#2 @henrikdvn
Regarding summary: I will try to make this more available to non-programmers and clarify that usage requires web programming skills.

Regarding the current state of the field: Some general aspects are fairly well described in the sources. but I guess most people don't look at them so I will provide some more details in the text. When it comes to developmental psychology, the main problem is that research experiments where infants engage actively with digital devices are quite rare. Preparing and engaging infants with limited linguistic capabilities is hard enough with familiar objects and interactions, but even harder with "virtual environments" presented on a tablet. Consequently, there is not much literature on how this is currently accomplished. However, I will have a second look.

Thanks for info on citation syntax. Will fix it :)
