### 2023-07-17 henrikdvn/CRSocket#1 @rabdill
Sounds good to me! I hadn't considered making the layout mobile-friendly, but that makes a lot of sense, as does your proposed solution. I only mention the community guidelines because it's part of the review criteria—I realize it's probably more involved than necessary for projects that don't have a big community working on it already, but it could at least be helpful to answer basic questions such as "How do I tell the authors about a bug I've encountered?" or "How should I submit a proposed change?" There are lots of guides and examples online for [codes of conduct](https://opensource.guide/code-of-conduct/) and ["contributor covenants,"](https://opensource.google/documentation/reference/releasing/template/CODE_OF_CONDUCT) but, at least from my perspective, the JOSS requirements don't call for anything so extensive.
