### 2023-07-17 henrikdvn/CRSocket#2 @rabdill
This sounds promising. If there are not many examples of this already being done, that may be another way to address this also—my impression, as someone in another field, was that this framework resolved issues that were unaddressed in competing products, but if this is the software that actually opens up new possibilities for working with a broader range of subjects, that's very exciting too.
