### 2023-07-24 #5658 @lukaszjablonski
I think it's OK if it is not tested on all available PHP versions and if it works with still supported 8.2.5 it should be fine.

@henrikdvn, maybe you could add note in the documentation on which versions of PHP it was tested?
