### 2023-07-27 aaddbd8
Update paper.md
The summary and Statement of need sections have been modified so that they are easier to understand for a non-specialist audience and more precise in terms of usability, functionality and research significance.
