+++ paper/paper.md
@@ -21,19 +21,18 @@ bibliography: paper.bib
 
 # Summary
 
-The Controller-Responder Socket is a modular, event-driven client/server web app infrastructure that can be used to develop digital tasks where testers use a controller device to administer trials and a separate responder device to present stimuli and collect responses. In addition to providing a more streamlined and comfortable test situation, this approach may improve measurement quality because potentially distracting activities related to configuration and initiation of tasks are hidden from the participant. It also facilitates more flexible data collection protocols where testers can make real-time adjustments based on direct observations and feedback from various sources presented on the controller device. In addition to communication functionality, it also provides client-side buffering of events and server-side plug-in interfaces related to access control and data storage.
-
+The Controller-Responder Socket (CRSocket) is a web app component that can be used to develop digital tasks where testers use a controller device to administer trials and a separate responder device to present stimuli and collect responses. In addition to providing a more streamlined and comfortable test situation, this approach may improve measurement quality because potentially distracting activities related to configuration and initiation of tasks are hidden from the participant. It also facilitates more flexible data collection protocols where testers can make real-time adjustments based on direct observations of the test situation and feedback from various sources presented on the controller device. In addition to communication functionality, the CRSocket also provides client-side buffering of events and server-side plug-in interfaces related to access control and data storage. Please note that usage of this software requires programming skills and experience with the PHP and JavaScript programming languages.
 # Statement of need
 
 The CRSocket component was originally developed for the Early Childhood Inhibitory Touch Task (ECITT). This is a suite of experimental psychology tasks designed to measure inhibitory behaviour using trials presented on tablet computers.
 
-The primary novelty of the ECITT is that it can be "minimally modified to suit different ages, whilst remaining structurally equivalent" [@Holmboe2021]. This is the first task that can be used to measure inhibitory behaviour in a consistent way across the lifespan, from 18 months and upward.  In addition to the longitudinal validation publication, three studies related to specific age groups based on this task, have been published so far [@Lui2021], [@Hendry2022], [@Fiske2022].
+The primary novelty of the ECITT is that it can be "minimally modified to suit different ages, whilst remaining structurally equivalent" [@Holmboe2021]. This is the first task that can be used to measure inhibitory behaviour in a consistent way across the lifespan, from 18 months and upward.  In addition to the longitudinal validation publication, three studies related to specific age groups based on this task, have been published so far [@Lui2021; @Hendry2022; @Fiske2022].
 
-Tablet computers and similar devices have become a natural part of daily life and are generally considered feasible for many types of data collection [@Frank2016], [@Semmelmann2016], [@Kanerva2019], but the abstract nature of such tasks still require careful consideration of potential sources of distraction [@Wenz2021] and adaption to individual differences among participants [@Jenkins2016]. Using digital tasks for developmental psychology experiments is particularly challenging. Infants are easily distracted and making them comfortable, motivated and capable of performing tasks is not trivial. In the ECITT this was accomplished by social interaction combined with simple games and priming trials on the responder device.
[…]
