### 2023-07-27 henrikdvn/CRSocket#2 @henrikdvn
I have now updated the paper and hopefully addressed most of these issues. Will appreciate all types of feedback.
