### 2023-08-01 998c976
Making the CommTest app more user friendly
