+++ CONTRIBUTING.md
@@ -1,0 +1,17 @@
+# Contributing guidelines
+
+## Contact
+
+I prefer to communicate by email. Please contact me at <henrik.dvergsdal@nord.no>
+
+## Contributing to the software
+
+I will consider incorporating modifications, extensions etc. that have been extensively tested.
+
+## Reporting issues
+
+I appreciate any type of feedback, especially issues related to usage in different browser engines, network connectivity conditions, and server environments.
+
+## Seeking support
+
+I will respoind to concrete questions about the software and use them as a basis for improving the documentation. 
