### 2023-08-02 henrikdvn/CRSocket#1 @henrikdvn
I have now made the CommTest app a bit more user-friendly by adding descriptive title attributes (shown when moving the cursor above headings etc.),  and adding info links to relevant documentation. Have also added some simple contribution guidelines and provided more precise version info in the Installation documentation.
