### 2023-08-02 #5658 @henrikdvn
> @henrikdvn, maybe you could add note in the documentation on which versions of PHP it was tested?

Done
