### 2023-08-02 #5658 @henrikdvn
@rabdill, @lukaszjablonski
Hopefully, I have now addressed all issues mentioned above
