### 2023-08-02 henrikdvn/CRSocket#3 @lukaszjablonski
JOSS Review: Functionality (Installation) and Installation instructions 
This is related to revision for JOSS openjournals/joss-reviews#5658.

I installed and successfully tested your application using [XAMPP 8.2.4 (PHP 8.2.4) for Windows](https://sourceforge.net/projects/xampp/files/XAMPP%20Windows/8.2.4/). Everything works fine for me but I would like to see a bit more details in installation instruction that would help less experienced users.

> Directories for logging, sockets, and data require write access.

Maybe it could help to write exact paths to this directories? Default paths are set in a `boot.php` so one could at least point out user to check it and set it already here.

Also, it would help to have above information as a comments in the `boot.php`.
