### 2023-08-02 #5658 @lukaszjablonski
Just not to open an issue for something that small, there is a typo in [CONTRIBURING.md](https://github.com/henrikdvn/CRSocket/blob/main/CONTRIBUTING.md): is "respoind", should be "respond".
