### 2023-08-02 henrikdvn/CRSocket#3 @henrikdvn
Good to hear that it works. Have added comments in each boot script and added an installation chapter describing these scripts which also includes links to currently assigned values in the project home page.
