### 2023-08-02 #5658 @lukaszjablonski
@henrikdvn: In [paper.md](https://github.com/henrikdvn/CRSocket/blob/b46868ef5f7e0d84878ff6c56bd33c2cdaa7c8db/paper/paper.md) Please change "eg." to "e.g." and "well defined set" to "well-defined set"
