### 2023-08-02 #5658 @lukaszjablonski
@henrikdvn: In [paper.md](https://github.com/henrikdvn/CRSocket/blob/b46868ef5f7e0d84878ff6c56bd33c2cdaa7c8db/paper/paper.md) I would also recommend checking consistency between terms in main text and figure: "CRSocket" vs "crSocket ", "CREvent" vs "crEvent", "clientid" vs "clientId". One could also try to use for those terms inline code quotes markup in main text but I don't know what is JOSS style guide here (@britta-wstnr, would ytou know?).
