### 2023-08-03 db489f3
Spelling error corrections and naming conventsions
