### 2023-08-03 #5658 @henrikdvn
Have fixed spelling mistakes, added a figure caption and included a section on basic naming conventions in the paper and on the Terminology page.
