### 2023-08-03 henrikdvn/CRSocket#2 @lukaszjablonski
I think, as already mentioned by @rabdill, the paper is still missing "State of the field". Current version of [paper.md](https://github.com/henrikdvn/CRSocket/blob/db489f3c688a4ac1cf770d353888246bf76b8a69/paper/paper.md) is explaining that others are employing tablets for the ECITT task:

> Current approaches to digital tasks are either based on self-administration or on handing over pre-configured devices to the participants.

As far as I understand from the paper, the difference of CRSocket to other available tools is that experimenter have better control since subject device can be controlled from experimenter device:

> Controlling tasks from a separate device can help meet some of the challenges mentioned above by enabling a bigger toolbox related to motivation, preparation and individual task adaption, based on researchers' ability to engage with and understand the participants.

This exact same idea and control is available in the setup described by Holmboe et al. (2021) to which the paper is referring already:

> The primary novelty of the ECITT is that it can be "minimally modified to suit different ages, whilst remaining structurally equivalent" [@Holmboe2021].

and the Holmboe et al. (2021) claim that subject device can be controlled by experimenter device too:

> Stimuli were presented on an Apple iPad tablet (the ‘responder’), with a screen size of 9.7 inches and a resolution of 2048 × 1536 pixels. The experimenter controlled stimulus presentation on the iPad via a smartphone (the ‘controller’) using a wi-fi network.

I would expect that as a "State of the field", for which JOSS ask question: "Do the authors describe how this software compares to other commonly-used packages?", the paper should provide a comparison or summary of what experimenter will gain by using CRSocket in comparison to tool used by  Holmboe et al. (2021) since as far as I understand both implement the ECITT.

Is  Holmboe et al. (2021) not available for public/free use? Also, are there other tools that implement the ECITT?

Such discussion will help reader and person potentially interested in using CRSockets to understand its benefits and will make the paper stronger.
