### 2023-08-03 #5658 @lukaszjablonski
For now, I have only two more points on my checklist to complete:
1. In a section "Documentation" - "A statement of need: Do the authors clearly state what problems the software is designed to solve and who the target audience is?" - I find such statement in the [README.md](https://github.com/henrikdvn/CRSocket/blob/6ca3b5df2e6d6ef100063e28eefff40db223f320/README.md) file of the repository. However, in documentation referenced in same [README.md](https://github.com/henrikdvn/CRSocket/blob/6ca3b5df2e6d6ef100063e28eefff40db223f320/README.md) (https://henrikdvn.github.io/CRSocket/ or https://neurolab.no/docs/crsocket/) there is no such statement. @britta-wstnr, what would be a take o JOSS? If that is OK with JOSS I'll be fine to check it out in my checklist.
2. In section "Software paper" - "State of the field: Do the authors describe how this software compares to other commonly-used packages?" -Although revision to the paper was made, I think that is still missing in the paper. I dropped more extensive comment on that in already closed issue (henrikdvn/CRSocket#2).
