### 2023-08-03 henrikdvn/CRSocket#2 @henrikdvn
Just to clarify: The software referenced in Holmboe et al. (2021) is the original ECITT software. The original code, which is an undocumented prototype with low reusability potential, was included with the paper as supplementary material. To my knowledge, there is currently no other tool implementing the ECITT test suite and no other software designed for digital tasks that implement controller-responder functionality. This is actually quite challenging. I have to turn down usage requests because I don't have time to support or maintain that software properly.  I hope, however, that the CRSocket component will make it easier to implement alternative tools for the ECITT and similar tasks.
