### 2023-08-03 henrikdvn/CRSocket#2 @lukaszjablonski
@henrikdvn, that is exactly what makes your software much more important here. I would highlight that in the paper and that was actually my only point.
