### 2023-08-03 0203da3
Added short description of functionality and puropose in the documentation
