+++ README.md
@@ -2,7 +2,7 @@
 
 ## Controller-Responder Socket
 
-The Controller-Responder Socket is a modular, event-driven client/server web app infrastructure that can be used to develop digital tasks where testers use a controller device to administer trials and a separate responder device to present stimuli and collect responses.
+The Controller-Responder Socket is a modular, event-driven client/server web app component that can be used to develop digital tasks where testers use a controller device to administer trials and a separate responder device to present stimuli and collect responses.
 
 Test apps and documentation can be found at the project home page: [neurolab.no/crsocket](https://neurolab.no/crsocket/)
 
