### 2023-08-03 #5658 @henrikdvn
Have included a short introduction of functionality and purpose in the documentation index page. Will try to address the second point tomorrow :)
