### 2023-08-04 e9aa2a0
Updated Statement of need and Architecture
