+++ paper/paper.md
@@ -31,19 +31,17 @@ The primary novelty of the ECITT is that it can be "minimally modified to suit d
 
 Tablet computers and similar devices have become a natural part of daily life and are generally considered feasible for data collection within various humanities disciplines [@Frank2016; @Semmelmann2016; @Kanerva2019], but the abstract nature of such tasks still require careful consideration and adaption to individual differences among participants and the context in which data is collected [@Wenz2021]. Using tablet tasks in developmental psychology research is challenging because the capacity to engage with stimuli presented on two-dimensional touch screens varies with age. Pre-verbal infants pose a challenge in most types of experimental research because they have to be introduced to tasks by means of social interaction and hands-on experiences rather than plain instructions. On tablet devices, this will typically involve different types of games and priming trials. Since infants are generally more distractible and impulsive than other age groups, it is particularly important to minimise irrelevant incidents and present tasks at the right moments. [@Jenkins2016] has identified similar challenges with elderly respondents. With this age group, one also has to take physical conditions such as visual acuity and motor skills into account.
 
-Current approaches to digital tasks are either based on self-administration or on handing over pre-configured devices to the participants. Controlling tasks from a separate device can help meet some of the challenges mentioned above by enabling a bigger toolbox related to motivation, preparation and individual task adaption, based on researchers' ability to engage with and understand the participants.
+To my knowledge, there is currently no other software designed for digital tasks that implement controller-responder functionality. Hence current approaches to tablet tasks are either based on self-administration or on handing over pre-configured devices to the participant. Controlling tasks from a separate device can help meet some of the challenges mentioned above by enabling a bigger toolbox related to motivation, preparation and individual task adaption, based on researchers' ability to engage with and understand the participants.
 
 Although the component was originally designed for lab experiments, it may also contribute to improved measurement quality and the creation of novel tablet tasks for other purposes, including settings where testers are not located in the same room but are able to communicate in other ways, e.g. via video channels.
 
 # Architecture
 
-To maximise reusability, the original software has been encapsulated in generic classes, using only built-in client- and server-side functionality and standard APIs. There are no global variables and no bindings to other parts of the ECITT software or external libraries. Client-side classes are implemented in JavaScript and server-side classes are implemented in PHP.
+To maximise reusability, the component has been extracted from the original software, an undocumented prototype included as supplementary material in [@Holmboe2021], and encapsulated in generic classes, using only built-in client- and server-side functionality and standard APIs. There are no global variables and no bindings to other parts of the ECITT software or external libraries. Client-side classes are implemented in JavaScript and server-side classes are implemented in PHP.
 
 The component implements a relatively robust protocol for the identification and exchange of events between controller and responder processes. It is designed to handle various technical and human incidents that has been observed during testing. This includes differences between different browser engines, network issues, and challenges related to server functionality, resource usage and delays.
 
 The protocol is asymmetric in the sense that only controller apps can establish and maintain communication channels. Responder apps can accept, reject, and abort connections, but they cannot actively approach controllers to establish new connections. Apart from that, controllers and responders interact as equal peers.
-
-Authentication is based on access tokens. This facilitates integration with a wide range of authentication schemes ranging from simple user ids to advanced industry-standard protocols.
 
 ## The CRSocket class
 
@@ -51,7 +49,7 @@ The CRSocket class is used by controller and responder client apps to establish 
 
[…]
