+++ paper/paper.md
@@ -15,7 +15,7 @@ authors:
 affiliations:
  - name: Nord University, Norway
    index: 1
-date: 16 May 2023
+date: 4 Aug 2023
 bibliography: paper.bib
 ---
 
@@ -69,7 +69,7 @@ Both eventsource and sendevent resources may contain server-side event plugin me
 
 
 ![Basic architecture](basicArchitecture.png)
-Basic architecture. Illustrates how CRSocket objects interact with server resources to obtain unique identifiers and exchange CREvent objects [^1].
+*Basic architecture. Illustrates how CRSocket objects interact with server resources to obtain unique identifiers and exchange CREvent objects [^1].*
 
 [^1]: Naming conventions: Classes and data types are named with medial capitals, starting with capital letters, e.g. "CRSocket".
 Class instances (objects) and variables are named with medial capitals, starting with lowercase letters, e.g. "crSocket".
