+++ paper/paper.md
@@ -67,17 +67,17 @@ Server-side functionality can be replaced by other implementations as long as th
 
 Both eventsource and sendevent resources may contain server-side event plugin mechanisms that can be used to implement authentication schemes, storage functionality and custom event processing. User authentication is based on access tokens. This facilitates integration with a wide range of authentication schemes ranging from simple user ids to advanced industry-standard protocols.
 
-
 ![Basic architecture](basicArchitecture.png)
+*Basic architecture. Illustrates how CRSocket objects interact with server resources to obtain unique identifiers and exchange CREvent objects.*
 
-*Basic architecture. Illustrates how CRSocket objects interact with server resources to obtain unique identifiers and exchange CREvent objects [^1].*
-
-[^1]: Naming conventions: Classes and data types are named with medial capitals, starting with capital letters, e.g. "CRSocket".
-Class instances (objects) and variables are named with medial capitals, starting with lowercase letters, e.g. "crSocket".
-Server resource are named with lowercase letters only, e.g. "eventsource".
+### Naming conventions:
+* Classes and data types are named with medial capitals, starting with capital letters, e.g. "CRSocket".
+* Class instances (objects) and variables are named with medial capitals, starting with lowercase letters, e.g. "crSocket".
+* Server resource are named with lowercase letters only, e.g. "eventsource".
 
 # Acknowledgements
 
 I would like to thank Andrew Simpson who came up with the original idea and Karla Holmboe who brought it into light by contributing with methodological expertise and integration into several developmental psychology research projects. I also appreciate the work of all the testers who provided useful feedback while patiently dealing with bugs and deficiencies.
+
 
 # References
