### 2023-08-04 #5658 @henrikdvn
Have added content related to currently available software, and software origin in the "Statement of need" and "Architecture" sections respectively.
