### 2023-08-04 #5658 @britta-wstnr
Awesome progress here @lukaszjablonski and @henrikdvn ! 🌱
Regarding your questions, @lukaszjablonski:
1. _Statement of need_: It looks like @henrikdvn has solved this by including text in the documentation pages? Can you confirm this addresses the point you were raising @lukaszjablonski ?
2. _Style guide_: I can see @henrikdvn added a section to the paper detailing naming conventions for the project. While not explicitly stated in the JOSS documentation, we do recommend using backticks for variable names or package names etc., see our example paper: https://joss.readthedocs.io/en/latest/submitting.html#example-paper-and-bibliography
