### 2023-08-04 #5658 @lukaszjablonski
From my side all points are checked and **the paper is ready for the publication in JOSS**.

In general, the author presents a well-documented working implementation of the Controller-Responder Socket (CRSocket). This implementation allows for creation of digital tasks where experimenter (in the paper called "testers") using a controller device administer trials on a separate device operated by experiment subject (in the paper called "responder"). As an example of CRSocket use, the author refers to Early Childhood Inhibitory Touch Task (ECITT) which is used in experimental psychology but its easy adaptation allows for creation of different tasks (controller/responder web app pairs).

CRSocket runs on a web server (online or local) that does not require a lot of resources. CRSocket client as a system-independent web-based application can be run on a variety of devices that run modern web browsers so that affordable devices could be employed. In my opinion, that makes CRSocket attractive for the number of users looking for tool that could implement tablet-based tasks.
