### 2023-08-12 #5658 @henrikdvn
So, @rabdill, what's your take on this paper?
