### 2023-08-16 #5658 @rabdill
Hi all, I'm grateful for your patience. I'm also delighted to say my checklist is complete, @henrikdvn did a great job responding to our feedback, and I don't have any objections to publication in JOSS. Please let me know if there's anything else I can do.
