### 2023-08-17 #5658 @britta-wstnr
and, for reference/bookkeeping since not all was tagged here, the issues related to this review:
https://github.com/henrikdvn/CRSocket/issues/1
https://github.com/henrikdvn/CRSocket/issues/2
https://github.com/henrikdvn/CRSocket/issues/3
