### 2023-08-22 #5658 @henrikdvn
@britta-wstnr Wasn't able to tick items on the list, but I have now:

Double-checked author info
Released v1.0.0
Archived the repository on figshare , doi: 10.6084/m9.figshare.24009225
Checked title, author info and license on figshare
