### 2023-08-25 #5658 @henrikdvn
@britta-wstnr: guess I should upload a final version to figshare once everything is complete.
