### 2023-08-30 #5658 @britta-wstnr
Hi @henrikdvn  I checked with a more experienced editor and there is no need to update the archive as that is predominantly meant for the software but not the paper. That being said, if you would want to update, you can of course do that and then let me know the new DOI.

If you could let me know what you prefer I can then take care of the rest of the checklist!
