### 2023-08-30 #5658 @henrikdvn
I have uploaded the current repository (with your updates) to figshare.

Just discovered that DOIs of new versions of figshare items consist of the original DOI with the version number appended. I had to update the title after the initial upload, so there were already two versions. This means that the DOI of this version is 10.6084/m9.figshare.24009225.v3 (https://doi.org/10.6084/m9.figshare.24009225.v3)

Seems like the original DOI always points to the latest version, but it's probably a good idea to include the version part anyway so that I can update it later, if needed, without compromising the publication DOI.
