### 2023-08-31 #5658 @editorialbot
:wave: @openjournals/sbcs-eics, this paper is ready to be accepted and published.
